delete testme file
