+++ test/unit/x31schemapack.test.js
@@ -412,6 +412,27 @@ describe('Openapi 3.1 schema pack validateTransactions', function() {
         .to.be.equal('The request body property \"\" should have required property \"notRequiredElement\"');
     });
   });
+
+  it('Should not generate any mismatch with a correct file using $schema', function() {
+    const collectionSource = path.join(__dirname, OPENAPI_31_COLLECTIONS + '/petstore$schemaCollection.json'),
+      collectionData = fs.readFileSync(collectionSource, 'utf8'),
+      schemaSource = path.join(__dirname, OPENAPI_31_COLLECTIONS + '/petstore$schema.json'),
+      schemaData = fs.readFileSync(schemaSource, 'utf8'),
+      validator = new SchemaPack({
+        type: 'string',
+        data: schemaData
+      });
+    let transactions = [];
+    getAllTransactions(JSON.parse(collectionData), transactions);
+
+    validator.validateTransaction(transactions, (err, result) => {
+      let requestIds = Object.keys(result.requests);
+      expect(err).to.be.null;
+      requestIds.forEach((requestId) => {
+        expect(result.requests[requestId].endpoints[0].matched).to.be.true;
+      });
+    });
+  });
 });
 
 describe('Openapi 3.1 schemapack mergeAndValidate', function() {
